requirements.txt
django==3.2.3
django-autoslug
#git+git://github.com/theatlantic/django-mysqlndb-backend.git@v1.1.0#egg=django-mysqlndb-backend
mysqlclient
